# docs/api/OPENAPI.yaml
openapi: 3.0.0
paths:
...
  /v1/societies/lookup:
    get:
      description:
        Resolves a join code to the society's public details. Returns name,
        city, state, type and member count only, and cannot be used to enumerate
        codes.
      operationId: SocietiesController_lookup
      parameters:
        - name: code
          required: true
          in: query
          description: The 6-character join code, case-insensitive.
          schema: {}
      responses:
        "200":
          description: The society the code belongs to.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      preview:
                        type: object
                        properties:
                          id:
                            type: string
                          name:
                            type: string
                          city:
                            type: string
                          state:
                            type: string
                          type:
                            type: string
                            enum:
                              - apartment
                              - villa
                              - rowhouse
                              - shared_flat
                              - other
                          memberCount:
                            type: integer
                        required:
                          - id
                          - name
                          - city
                          - state
                          - type
                          - memberCount
                        additionalProperties: false
                    required:
                      - preview
                    additionalProperties: false
                  meta:
                    type: object
                    properties:
                      requestId:
                        type: string
                      timestamp:
                        type: string
                        format: date-time
                    required:
                      - requestId
                      - timestamp
                    additionalProperties: false
                required:
                  - data
                  - meta
                additionalProperties: false
        "400": *a8
        "401": *a9
        "403": *a10
        "404": *a11
        "409": *a12
        "422": *a13
        "429": *a14
      security: *a15
      summary: Preview a join code (public)
      tags: *a16
  /v1/societies/join:
    post:
      description:
        "Resolves the code, checks expiry against the server clock and
        records a membership. Never auto-approved: the membership comes back
        `pending` until an Admin approves it (PRD §3.2)."
      operationId: SocietiesController_join
      parameters: []
      responses:
        "201":
          description: The membership that was created, or asked again.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      membership:
                        type: object
                        properties:
                          id:
                            type: string
                          societyId:
                            type: string
                          userId:
                            type: string
                          role:
                            type: string
                            enum:
                              - admin
                              - treasurer
                              - committee_member
                              - resident
                              - tenant
                              - guest
                          status:
                            type: string
                            enum:
                              - pending
                              - active
                              - removed
                          occupancyType:
                            type: string
                            enum:
                              - owner
                              - tenant
                              - family_member
                          joinedAt:
                            nullable: true
                            type: string
                        required:
                          - id
                          - societyId
                          - userId
                          - role
                          - status
                          - occupancyType
                          - joinedAt
                        additionalProperties: false
                    required:
                      - membership
                    additionalProperties: false
                  meta:
                    type: object
                    properties:
                      requestId:
                        type: string
                      timestamp:
                        type: string
                        format: date-time
                    required:
                      - requestId
                      - timestamp
                    additionalProperties: false
                required:
                  - data
                  - meta
                additionalProperties: false
        "400": *a8
        "401": *a9
        "403": *a10
        "404": *a11
        "409": *a12
        "422": *a13
        "429": *a14
      security: *a15
      summary: Join a society with a code
      tags: *a16
...
  /v1/societies/{societyId}/join-code:
    post:
      description:
        Admin-only. The new code is minted server-side and invalidates the
        previous one immediately, so a leaked code is contained by one action.
      operationId: SocietiesController_regenerateJoinCode
      parameters:
        - name: societyId
          required: true
          in: path
          description: Society UUID.
          schema:
            type: string
      responses:
        "200":
          description: The society, with its new join code.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      society:
                        type: object
                        properties:
                          id:
                            type: string
                          name:
                            type: string
                          slug:
                            type: string
                          type:
                            type: string
                            enum:
                              - apartment
                              - villa
                              - rowhouse
                              - shared_flat
                              - other
                          registrationNumber:
                            nullable: true
                            type: string
                          addressLine1:
                            nullable: true
                            type: string
                          addressLine2:
                            nullable: true
                            type: string
                          city:
                            type: string
                          state:
                            type: string
                          pincode:
                            nullable: true
                            type: string
                          country:
                            type: string
                            enum:
                              - IN
                          currency:
                            type: string
                            enum:
                              - INR
                          timezone:
                            type: string
                          joinCode:
                            type: string
                          joinCodeExpiresAt:
                            nullable: true
                            type: string
                          plan:
                            type: string
                            enum:
                              - free
                              - pro
                              - enterprise
                          createdBy:
                            type: string
                          createdAt:
                            type: string
                          updatedAt:
                            type: string
                          deletedAt:
                            nullable: true
                            type: string
                          settings:
                            type: object
                            properties:
                              billingDay:
                                type: integer
                              dueDay:
                                type: integer
                              graceDays:
                                type: integer
                              approvalThresholdPaise:
                                type: integer
                              billVacantFlats:
                                type: boolean
                              allowPartialPayments:
                                type: boolean
                              defaulterListPublic:
                                type: boolean
                              financialYearStartMonth:
                                type: integer
                              timezone:
                                type: string
                              currency:
                                type: string
                                enum:
                                  - INR
                            required:
                              - billingDay
                              - dueDay
                              - graceDays
                              - approvalThresholdPaise
                              - billVacantFlats
                              - allowPartialPayments
                              - defaulterListPublic
                              - financialYearStartMonth
                              - timezone
                              - currency
                            additionalProperties: false
                          memberCount:
                            type: integer
                        required:
                          - id
                          - name
                          - slug
                          - type
                          - registrationNumber
                          - addressLine1
                          - addressLine2
                          - city
                          - state
                          - pincode
                          - country
                          - currency
                          - timezone
                          - joinCode
                          - joinCodeExpiresAt
                          - plan
                          - createdBy
                          - createdAt
                          - updatedAt
                          - deletedAt
                          - settings
                          - memberCount
                        additionalProperties: false
                    required:
                      - society
                    additionalProperties: false
                  meta:
                    type: object
                    properties:
                      requestId:
                        type: string
                      timestamp:
                        type: string
                        format: date-time
                    required:
                      - requestId
                      - timestamp
                    additionalProperties: false
                required:
                  - data
                  - meta
                additionalProperties: false
        "400": *a8
        "401": *a9
        "403": *a10
        "404": *a11
        "409": *a12
        "422": *a13
        "429": *a14
      security: *a15
      summary: Regenerate the join code
      tags: *a16
  /v1/societies/{societyId}/leave:
    post:
      description: Leaves, or withdraws a pending request. Refused with
        `SOCIETY_ADMIN_REQUIRED` if the caller is the only active Admin — a
        society can never be left without one.
      operationId: SocietiesController_leave
      parameters:
        - name: societyId
          required: true
          in: path
          description: Society UUID.
          schema:
            type: string
      responses:
        "204":
          description: Left. The caller's membership is now `removed`.
        "400": *a8
        "401": *a9
        "403": *a10
        "404": *a11
        "409": *a12
        "422": *a13
        "429": *a14
      security: *a15
      summary: Leave a society
      tags: *a16
info:
  title: Society Expense Splitter API
  description: "The SES API. Money is always integer paise in a `*Paise` field;
    timestamps are ISO-8601 UTC with milliseconds. Success responses are `{
    data, meta }`; failures are `{ error: { code, message, requestId, timestamp,
    docs } }`."
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    supabase-jwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
...